# docs/openapi/famlin.yaml
...
openapi: 3.0.3
info:
  title: Famlin API
  version: current
  description: |
    Every Famlin server exposes this REST API under `/api` — the same API the
    official mobile and web apps use. With a [personal access token](/developers/api)
    you can build your own integrations against your family's server.

    All content is scoped by **group membership**: you can only read and write
    content in groups (families) you belong to, and there is no way around that
    per token. See [Using the API](/developers/api) for authentication,
    conventions, and worked examples.
servers:
  - url: https://famlin.example.com
    description: Your own Famlin deployment
security:
  - bearerAuth: []
...
paths:
...
  /api/api-tokens:
    get:
      operationId: listApiTokens
      tags: [API tokens]
      summary: List your API tokens
      description: Never contains token secrets — only the display preview.
      responses:
        '200':
          description: Your tokens, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiToken'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createApiToken
      tags: [API tokens]
      summary: Create an API token
      description: |
        Returns the plaintext secret **exactly once** — it is stored only as a
        SHA-256 hash and can never be retrieved again.

        Requires a **session-authenticated** request: calling this with another
        API token returns `403`, so a leaked token can't create replacements
        for itself and outlive its own revocation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: Photo frame script
                expiresInDays:
                  type: integer
                  minimum: 1
                  maximum: 3650
                  description: Omit for a token that never expires.
      responses:
        '201':
          description: The created token, including its secret (shown only here)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedApiToken'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: The request was authenticated with an API token instead of a login session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/api-tokens/{id}:
    delete:
      operationId: revokeApiToken
      tags: [API tokens]
      summary: Revoke an API token
      description: Takes effect immediately. You can only revoke your own tokens.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Not found (or not your token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/groups:
    get:
      operationId: listGroups
      tags: [Groups]
      summary: List your groups
      responses:
        '200':
          description: Groups you belong to
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/posts:
    get:
      operationId: listPosts
      tags: [Posts]
      summary: List posts (your feed)
      description: |
        Newest first, cursor-paginated. The feed is a filter over your groups:
        `groupIds` selects a subset, the legacy `groupId` selects one, and
        neither means every group you belong to. Requesting a group you're not
        a member of is a `403`.
      parameters:
        - name: groupIds
          in: query
          schema: { type: string }
          description: Comma-separated group ids (must all be groups you're in).
          example: cmb1ghi,cmb1jkl
        - name: groupId
          in: query
          schema: { type: string }
          description: Legacy single-group filter.
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/take'
      responses:
        '200':
          description: A page of posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostsPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/NotGroupMember'
    post:
      operationId: createPost
      tags: [Posts]
      summary: Create a post
      description: |
        `uploadedAssetUrls` must be paths returned by the
        [upload endpoint](/developers/api-reference/upload-files) (`/uploads/...`)
        or Immich proxy paths for an album linked to **this** post's group —
        arbitrary URLs are rejected. `latitude` and `longitude` must be
        provided together.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [groupId]
              properties:
                groupId: { type: string }
                content:
                  type: string
                  maxLength: 5000
                type:
                  type: string
                  enum: [UPDATE, MILESTONE]
                  default: UPDATE
                milestoneTag:
                  type: string
                  maxLength: 50
                  example: First steps!
                uploadedAssetUrls:
                  type: array
                  maxItems: 20
                  items: { type: string }
                  example: ['/uploads/8d0f3c9a-1b2e-4f5a-9c8d-7e6f5a4b3c2d.jpg']
                latitude:
                  type: number
                  minimum: -90
                  maximum: 90
                longitude:
                  type: number
                  minimum: -180
                  maximum: 180
                locationName:
                  type: string
                  maxLength: 200
      responses:
        '200':
          description: The created post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/NotGroupMember'
...
  /api/notifications:
    get:
      operationId: listNotifications
      tags: [Notifications]
      summary: List your notifications
      description: The 100 most recent notifications, newest first.
      responses:
        '200':
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/uploads:
    post:
      operationId: uploadFiles
      tags: [Uploads & media]
      summary: Upload photos or videos
      description: |
        Allowed: jpg, jpeg, png, gif, webp, heic, heif, mp4, mov, m4v, webm.
        Max 200 MB per file, 10 files per request, 60 uploads per 10 minutes.
        Pass the returned paths into `uploadedAssetUrls` when creating a post,
        or into `avatarUrl` when updating your profile.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: The stored paths, in upload order
          content:
            application/json:
              schema:
                type: object
                properties:
                  urls:
                    type: array
                    items: { type: string }
                    example: ['/uploads/8d0f3c9a-1b2e-4f5a-9c8d-7e6f5a4b3c2d.jpg']
        '400':
          description: Unsupported file type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/push-tokens:
    post:
      operationId: registerPushToken
      tags: [Push tokens]
      summary: Register an Expo push token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      operationId: unregisterPushToken
      tags: [Push tokens]
      summary: Unregister an Expo push token
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Unregistered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A personal access token (`famlin_pat_...`, created in the web app) or
        a session JWT from `POST /api/auth/login`. Both are sent the same way:
        `Authorization: Bearer <token>`.
  parameters:
    cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Cursor from the previous page's `nextCursor`.
    take:
      name: take
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 30
      description: Page size.
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or revoked credential
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotGroupMember:
      description: You are not a member of this group
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: Invalid request body or query (see `details`)
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  details:
                    type: object
                    description: Zod field errors.
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable, translated via the `Accept-Language` header (`en`/`nl`).
      required: [error]
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ReactionType:
      type: string
      enum: [LIKE, LOVE, HAHA, WOW, SAD, CARE]
      default: LIKE
    ReactionCounts:
      type: object
      description: Per-emoji reaction counts; only emojis with at least one reaction appear.
      additionalProperties:
        type: integer
      example:
        LOVE: 3
        LIKE: 2
...
    UserSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        avatarUrl:
          type: string
          nullable: true
...
    Group:
      type: object
      properties:
        id: { type: string }
        name: { type: string, example: Family Smith }
        description:
          type: string
          nullable: true
        createdAt: { type: string, format: date-time }
        joinedAt:
          type: string
          format: date-time
          description: When *you* joined this group.
...
    Post:
      type: object
      properties:
        id: { type: string }
        authorId: { type: string }
        author:
          $ref: '#/components/schemas/UserSummary'
        groupId: { type: string }
        group:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
        content:
          type: string
          nullable: true
          example: First day of school!
        type:
          type: string
          enum: [UPDATE, MILESTONE]
        milestoneTag:
          type: string
          nullable: true
        uploadedAssetUrls:
          type: array
          items: { type: string }
          description: '`/uploads/...` or `/api/immich/assets/...` paths — auth-gated, see Uploads & media.'
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        locationName:
          type: string
          nullable: true
        createdAt: { type: string, format: date-time }
        editedAt:
          type: string
          format: date-time
          nullable: true
        commentCount: { type: integer }
        likeCount: { type: integer }
        likedByMe: { type: boolean }
        myReaction:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ReactionType'
        reactions:
          $ref: '#/components/schemas/ReactionCounts'
        recentReactors:
          type: array
          maxItems: 3
          items:
            $ref: '#/components/schemas/UserSummary'
          description: The three most recent reactors, newest first.
        favoritedByMe: { type: boolean }
    PostsPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        nextCursor:
          type: string
          nullable: true
          description: Pass back as `?cursor=` for the next page; `null` on the last page.
...
    ApiToken:
      type: object
      properties:
        id: { type: string }
        name: { type: string, example: Photo frame script }
        tokenPreview:
          type: string
          description: First characters of the secret, for telling tokens apart.
          example: Zx9kQ2aB
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
          description: Updated at most every few minutes while the token is in use.
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt: { type: string, format: date-time }
    CreatedApiToken:
      allOf:
        - $ref: '#/components/schemas/ApiToken'
        - type: object
          properties:
            token:
              type: string
              description: The full secret — shown only in this response.
              example: famlin_pat_Zx9kQ2aB7cD4eF6gH8iJ0kL1mN3oP5qR7sT9uV1wX3y
    Notification:
      type: object
      properties:
        id: { type: string }
        type:
          type: string
          description: '`new_post`, `new_comment`, `new_like_post`, `new_like_comment`, `mention`, or `on_this_day`.'
        message: { type: string }
        relatedPostId:
          type: string
          nullable: true
        post:
          type: object
          nullable: true
          properties:
            id: { type: string }
            groupId: { type: string }
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt: { type: string, format: date-time }
...